Add option setting and configuration write
Add a number of functions from the original API, such as `AddSection`,
`HasSection`, `HasOption`, `Set`, `Write`, `RemoveOption` and
`RemoveSection`.

Note: there are intentional differences in this implementation compared
to the original Python API: `RemoveOption` and `RemoveSection` return an
error to be more idiomatic and consistent with other Go calls, and the
additional convenience function `WriteFile` was added.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>